refactor contexts
